r036: bug fix and new heuristics
Added a routine to handle 2-target tips. Given A->B and A->C (A is a tip),
it might be desirable to ignore A and link B & C if A is a short tip, B & C
is not a pair of haplotigs, B & C used to have overlaps. Especially for
the cases where B and C are long with high coverage, and A is short with
much a lower coverage (*not in the actual criteria tho) as they seem to
be broken haplotype / confusion caused by incomplete haplotype info.